HBSD: llvm12 doesn't like to have its static libs built with llvm-ar
HardenedBSD switched to llvm-ar (for CFI) years ago. It turns out that
llvm12's internal static libraries don't like to be created using
llvm-ar as ar/ranlib.

I've emailed the llvm-dev mailing list and am waiting to hear back. For
now, make llvm12 use ar/ranlib from the elftoolchain project in base.

The good news is that if we get this fixed for building llvm12, the fix
will likely help with all the other ports broken with llvm-ar.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
issue:		#3