HBSD: Tell the compiler to ignore unused arguments
Some ports dislike our hardening CFLAGs. Since they don't use the
arguments we pass to the compiler, they fail, erroring on the unused
argument (the hardening CFLAGs).

For non-gcc ports, pass in -Qunused-arguments in order to silence those
failures. The -Qunused-arguments argument is not available for ports
that use the GNU compiler toolchain.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
issue:		#10