HBSD: Fix ports-mgmt/pkg
A bad commit[0] by FreeBSD caused `pkg repo` to fail. In their commit
message, they fail to give the reasoning behind the change, so I'm left
to guessing. My guess is that it's just broken.

Thanks to those HardenedBSD users who helped report, reproduce, and
test. Since the FreeBSD project didn't announce to the wider FreeBSD
community some of the more substantial changes to pkg (from 1.16.3 to
1.17.1), the HardenedBSD project was caught unaware of those changes.

It's my hope to learn from this in that communication is key. Projects
who have downstream consumers need to be aware of major changes well in
advance of landing (and releasing) the change. Downstream consumers
absolutely need time to plan to adopt these changes.

Another key takeaway is that for certain FreeBSD projects, following
eacn and every commit is necessary. My (already strained) time must be
used in auditing each commit of the FreeBSD projects HardenedBSD relies
on. This is especially even more important given FreeBSD's lack of
communication.

[0]: https://github.com/freebsd/pkg/commit/727f1c948eb5fe2d16c867db9e195c50166bf903

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>