HBSD: Enable SMB in multimedia/kodi
According to my tests, the SMB protocol support is particularly good in Kodi.

After contacting the port maintainer (yzrh),
he does not want to activate it by default because he does not use it (he prefers NFS).
I prefer to activate it in order to have a functional package, Ubuntu, Fedora
& Cie have functional packages, so why not Freebsd/HardenedBSD?

Signed-off-by: Loic <loic.f@hardenedbsd.org>