HBSD: Disable DTrace support for many ports
The llvm compiler toolchain is a bit strict, causing ports that support
DTrace by default to fail to build.

As much as I hate to do so, we need to disable DTrace support for these
ports until I get the time to file a proper bug report in upstream llvm.
Once we get a fix in llvm, we can revert this commit (which is the
ultimate goal).

An example build failure is at [0].

[0]: http://ci-03.md.hardenedbsd.org/data/hardenedbsd-13_amd64-local/2021-12-08_08h51m35s/logs/errors/tcl86-8.6.12.log

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	BlackhawkNest, Inc