HBSD: force the construction of lang/gnustep-base
Removed the use of USE_HARDENING because we must have the same
configuration as devel/gnustep-make this will remove the following error:
'The Objective-C compiler does not work or is not installed properly.'

Added missing dependency for libffcall.

Added a 'patch-configure' to work around the following problem:
pid 85185 (conftest), jid 0, uid 0: exited on signal 11 (core dumped)

Use of the "--disable-invocations" argument to bypass the use
of "--enable-native-objc-exceptions" in devel/gnustep-make.

Signed-off-by: Loic <loic.f@hardenedbsd.org>