HBSD: Remove auto-application of STACKAUTOINIT hardening feature
Now that HardenedBSD 14-CURRENT (amd64 and arm64) and 13-STABLE natively
apply -ftrivial-auto-var-init=zero, we don't need to tightly integrate
with it in the ports hardening framework. The compiler toolchain in base
simply does the right thing (applying it on behalf of ports).

We still provide a USES flag so that ports can effectively opt-out of
the trivial variable auto init feature. To opt a port out, add this USES
line:

```
USES+=	stack_autoinit:unintialized
```

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>