HBSD: Fix llvm compilter toolchain tests
The llvm compiler toolchain components have been installed as hardlinks
rather than symlinks for a while. I had forgotten to reflect that change
in ports, causing some hardening features to be force disabled, even for
ports that have those features enabled (example: CFI for
security/openvpn).

Due to the components being installed as hardlinks, we need to execute
the components in order to perform version checks. All llvm compiler
toolchain components report LLVM in their version output. Executing the
tools on every port build is sub-optimal, but it's what we need to do
for the moment.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>