HBSD: Restore libcurl's ability to resolve .onion domains
Earlier this year, the curl project decided to prevent working within a
human rights-centric environment. HardenedBSD provides Tor Onion Service
endpoints for all parts of its public infrastructure. With the world
standardizing over the libcurl monoculture just as it has with the
OpenSSL monoculture, it is critical that we undo this absurdity.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Fixes:		https://github.com/curl/curl/commit/0ae0abbe72514a75c10bfc4108d9f254f594c086