HBSD: Disable SIMD in audio/pulseaudio
This "fixes" the build of audio/pulseaudio by disabling SIMD support.
Pulseaudio's build framework misdetects which SIMD implementation to
use. Because HardenedBSD ships with a cross-compiler, Pulseaudio's build
framework thinks that since it was able to compile (not execute) the ARM
NEON based implementation, it should try to use that at link time, even
if we're on amd64.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Issue:		#34